Issue mcp-graveyard/UXP#1251 - Remove obsolete XP_QSORT reference.
I didn't include this in my initial PR because I assumed this part
was not relevant to UXP. I was asked to submit this while the MailNews rebase
was going on, but I was late with my patch and didn't quite grasp that this
specifically was still needed.

My counterpart at Mozilla, Petr Sumbara, has gotten his version of this
approved.

https://bugzilla.mozilla.org/show_bug.cgi?id=1568473